refactor: card와 participant 관련 패키지 생성 및 이동
