+++ src/main/java/blackjack/domain/participant/ParticipantName.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 public record ParticipantName(
         String name
