+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 public class Player extends Participant {
