+++ src/main/java/blackjack/domain/participant/Players.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.deck.Deck;
 import java.util.List;
 
 public class Players {
