+++ src/test/java/blackjack/domain/deck/DeckTest.java
@@ -1,7 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.Card;
 import org.junit.jupiter.api.Test;
 
 public class DeckTest {
